Welcome: Use arrow function for click handler
The onClick property in <button> expects a function that it will call on
click. Apparently, if you give it a function defined like this,

  click() {...}

it doesn't work. But it does if you define it like this:

  click = () => {...}

The reason is that the first one doesn't have a valid `this` reference
unless you explicitly give it one via binding-- e.g.

  this.click = this.click.bind(this)

Though you can still do the first one if you wrap it, which is what I
had:

  onclick={() => this.click()}

Ref: https://www.youtube.com/watch?v=kVWpBtRjkCk