server: Add script to fetch data from SDOT
This commit adds downloader.py, a Python script that facilitates
downloading SDOT data. The script fetches two datasets: 'Collisions All
Years' and 'Persons'. The first has an entry for every collision; the
second has an entry for every person involved in a collision (it's much
larger than the first).

It seems that the site has to prepare the data, since the first few
requests result in a response that says "Pending" or "Exporting Data".
The script polls the site until the download actually happens.

Oh and also recently, the request for 'Collisions All Years' always
results in "Pending"-- there might be a server-side problem there.

Downloaded data is thrown into the data/ folder. Later, the plan is to
throw it into sqllite tables.