Add app.py with basic routes backed by db queries
This commit adds app.py which, when run, starts up a server using the
Flask library. The server has some basic routes that query the database
created by db/load.py.

Also included are model files for both tables in the database--
'collision' and 'person'. These were created by first generating them
with sqlacodegen (code not committed) and then hand-editing them to the
ORM format. The SQLAlchemy unified tutorial was pretty helpful for
making the model fiels and doing the simple queries.

Tested the routes with Postman and got JSON back as expected.