client: Fetch collisions data from backend
This commit updates the simple client/frontend we have to fetch
collisions data from the backend instead of the mock file we have.

The query implementing the /collisions/many endpoint was also updated
to only return rows with non-NULL coordinate values. Maybe we will want
to clean these out or deal with them in some way.