db/load.py: Change file_path from attribute to method
Small refactor to make file_path a method. All it did was return the
name attribute surrounded by static strings to make it a full file path.

This was preventing me to regenerate the db without downloading the
files again. With this change, all it takes to do that now is to comment
out line 50: `self._download_data_from_sdot()`. (Eventually will make
this a CLI option.)