Add new 'intersection' table and supporting files
This commit a new script, db/load-intersection.py, for generating the
new 'intersection' table. The intersection table contains an entry for
every distinct location in the 'collisions' table. Each entry has
aggregate numbers for all the collisions that happened there.

(Now remembering that locations can also be mid-block, so perhaps
'intersection' was not the best name).

Also added:
- A model file defining the columns in the new table.
- A simple route that returns all intersections from the table.

Part of the motivation behind this was to make the heatmap much easier
to load. Instead of plotting each collision independently (~250K), we
now only need to plot every intersection (~26K), and weigh the "heat"
based on the aggregates at each one.

One more thing: as a compromise with mypy, we had to prepend the local
imports in db/load-intersection.py (models.collision) with `db.`. The
result is that we can't run the script directly (`python3 ...`); it must
run it as a module: `python3 -m db.load-intersection`. :/